Avatar par default

+++ src/UserBundle/EventSubscriber/AvatarUploadSubscriber.php
@@ -37,11 +37,13 @@ class AvatarUploadSubscriber implements EventSubscriber
         $entity = $event->getEntity();
         if ($entity instanceof Avatar) {
             $file = $entity->getFile();
-            $fileName = $entity->getName().'.'.$entity->getExtension();
-            $file->move(
-                $this->projectDir.'/web/upload',
-                $fileName
-            );
+            if(!is_null($file)){
+                $fileName = $entity->getName().'.'.$entity->getExtension();
+                $file->move(
+                    $this->projectDir.'/web/upload',
+                    $fileName
+                );
+            }
         }
     }
 
@@ -62,7 +64,7 @@ class AvatarUploadSubscriber implements EventSubscriber
     {
         $entity = $event->getEntity();
         if ($entity instanceof Avatar) {
-            if(file_exists($this->projectDir.'/web/upload/'.$entity->getOldFileName())){
+            if(!is_null($entity->getOldFileName()) && file_exists($this->projectDir.'/web/upload/'.$entity->getOldFileName())){
                 unlink ( $this->projectDir.'/web/upload/'.$entity->getOldFileName());
             }
         }
